backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
pymongo
redis
aio_pika
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
pydantic-settings
email-validator
motor
aio-pika
alembic
cloudinary
python-multipart
stripe
